refactor：系统定时任务采用动态配置、系统定时不支持删除、修复pear更新样式问题、修复消息中心列表

+++ src/main/java/com/tangerineSpecter/oms/system/service/system/impl/ScheduledManageServiceImpl.java
@@ -3,15 +3,19 @@ package com.tangerinespecter.oms.system.service.system.impl;
 import cn.hutool.core.convert.Convert;
 import com.tangerinespecter.oms.common.constants.RetCode;
 import com.tangerinespecter.oms.common.enums.GlobalBoolEnum;
+import com.tangerinespecter.oms.common.enums.ScheduledTypeEnum;
 import com.tangerinespecter.oms.common.exception.BusinessException;
 import com.tangerinespecter.oms.common.mapper.QueryWrapperX;
 import com.tangerinespecter.oms.common.query.SystemScheduledQueryObject;
+import com.tangerinespecter.oms.common.redis.KeyPrefix;
+import com.tangerinespecter.oms.common.redis.RedisKey;
 import com.tangerinespecter.oms.job.schedule.AbstractJob;
 import com.tangerinespecter.oms.system.convert.system.ScheduledConvert;
 import com.tangerinespecter.oms.system.domain.entity.SystemScheduledTask;
 import com.tangerinespecter.oms.system.domain.vo.base.IdParamVo;
 import com.tangerinespecter.oms.system.domain.vo.system.SystemScheduledVo;
 import com.tangerinespecter.oms.system.mapper.SystemScheduledTaskMapper;
+import com.tangerinespecter.oms.system.service.helper.RedisHelper;
 import com.tangerinespecter.oms.system.service.system.IScheduledManageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,6 +23,7 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @author TangerineSpecter
@@ -31,10 +36,13 @@ public class ScheduledManageServiceImpl implements IScheduledManageService {
 
     private final SystemScheduledTaskMapper scheduledTaskMapper;
     private final ApplicationContext context;
+    private final RedisHelper redisHelper;
 
     @Override
     public List<SystemScheduledTask> list(SystemScheduledQueryObject qo) {
-        return scheduledTaskMapper.selectList(new QueryWrapperX<SystemScheduledTask>().likeIfPresent("name", qo.getName()).eq("is_del", GlobalBoolEnum.FALSE.getValue()));
+        return scheduledTaskMapper.selectList(new QueryWrapperX<SystemScheduledTask>()
+                .likeIfPresent("name", qo.getName()).eqIfPresent("type", qo.getType())
+                .eq("is_del", GlobalBoolEnum.FALSE.getValue()));
     }
 
     /**
@@ -99,12 +107,21 @@ public class ScheduledManageServiceImpl implements IScheduledManageService {
     @Override
     public void delete(Long id) {
         SystemScheduledTask task = scheduledTaskMapper.selectById(id);
+        if (Objects.equals(ScheduledTypeEnum.DEFAULT.getValue(), task.getType())) {
+            throw new BusinessException(RetCode.DEFAULT_TASK_NOT_DELETE);
+        }
         this.cancelJob(task.getClassPath());
         scheduledTaskMapper.deleteById(id);
     }
 
     @Override
     public void executeJob(IdParamVo param) {
+        final RedisKey redisKey = RedisKey.getJobLock;
+
+        if (!redisHelper.lock(redisKey, param.getId(), "lock")) {
+            throw new BusinessException(RetCode.TASK_EXECUTE_RUNNING);
+        }
+
         try {
             SystemScheduledTask systemScheduledTask = scheduledTaskMapper.selectById(param.getId());
             AbstractJob job = (AbstractJob) context.getBean(Class.forName(systemScheduledTask.getClassPath()));
@@ -117,4 +134,5 @@ public class ScheduledManageServiceImpl implements IScheduledManageService {
             throw new BusinessException(RetCode.TASK_EXECUTE_ERROR);
         }
     }
+
 }
